refactor: move ExoPlayerBitrateLogger out of ExoPlayerPlayback
